fix npe

+++ integration-test/src/main/java/org/apache/iotdb/it/env/cluster/node/AbstractNodeWrapper.java
@@ -488,6 +488,9 @@ public abstract class AbstractNodeWrapper implements BaseNodeWrapper {
 
   @Override
   public void stopForcibly() {
+    if (this.instance == null) {
+      return;
+    }
     try {
       this.instance.destroyForcibly().waitFor(10, TimeUnit.SECONDS);
     } catch (InterruptedException e) {
